@@ -14,20 +14,23 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  firebase_auth: ^4.4.1
-  firebase_core: ^2.9.0
-  cloud_firestore: ^4.5.1
   multi_select_flutter: ^4.1.2
   intl: ^0.17.0
   bloc: ^8.0.3
   flutter_bloc: ^8.0.1
   flutter_localizations:
     sdk: flutter
+  dio: ^4.0.6
+  eventsource: ^0.4.0
+
+  flutlab_logcat:
+    git: https://github.com/FlutLab/flutlab-logcat.git
+
 
 dev_dependencies:
   flutter_test:
