@@ -1,8 +1,15 @@
 class Usuario {
-  final String uid;
-  final String email;
-  final String nomeLanchonete;
-  final String endereco;
+  late final String uid;
+  late final String email;
+  late final String nomeLanchonete;
+  late final String endereco;
 
   Usuario(this.uid, this.email, this.nomeLanchonete, this.endereco);
+
+  Usuario.fromMap(dados) {
+    uid = dados['id'];
+    email = dados['email'];
+    nomeLanchonete = dados['nomeLanchonete'];
+    endereco = dados['endereco'];
+  }
 }
