@@ -4,14 +4,18 @@ import 'package:bloc/bloc.dart';
 import 'package:lanchonete/cardapio/bloc/produto_cardapio_event.dart';
 import 'package:lanchonete/cardapio/bloc/produto_cardapio_state.dart';
 import 'package:lanchonete/cardapio/models/produto_cardapio.dart';
-import 'package:lanchonete/Provider/cardapio_provider.dart';
+import 'package:lanchonete/provider/cardapio_provider.dart';
+import 'package:lanchonete/provider/stream_provider.dart';
+import 'package:lanchonete/stream/bloc/stream_bloc.dart';
+import 'package:lanchonete/stream/bloc/stream_event.dart';
+import 'package:lanchonete/stream/controller/controlador_stream_api.dart';
 
-class ProdutoCardapioBloc extends Bloc<ProdutoCardapioEvent, ProdutoCardapioState> {
-  StreamSubscription? _listener;
+class ProdutoCardapioBloc
+    extends StreamBloc<ProdutoCardapioEvent, ProdutoCardapioState> {
+  final CardapioApiProvider _cardapioProvider = CardapioApiProvider.helper;
 
-  ProdutoCardapioBloc() : super(CarregandoProdutosCardapio()) {
-    on<IniciarStreamProdutosCardapio>(_iniciarStream);
-    on<PararStreamProdutosCardapio>(_pararStream);
+  ProdutoCardapioBloc()
+      : super(CardapioApiProvider.helper, CarregandoProdutosCardapio()) {
     on<CarregarProdutosCardapio>(_carregarProdutos);
     on<CarregarProdutoCardapio>(_carregarProduto);
     on<AdicionarAoCardapio>(_adicionarProduto);
@@ -19,64 +23,77 @@ class ProdutoCardapioBloc extends Bloc<ProdutoCardapioEvent, ProdutoCardapioStat
     on<RemoverDoCardapio>(_removerProduto);
   }
 
-  void _iniciarStream(IniciarStreamProdutosCardapio event, Emitter<ProdutoCardapioState> emit) {
-    _listener = CardapioFirestoreServer.helper.iniciarStreamProdutosCardapio(event.uid).listen((event) {
-      ProdutoCardapioState atualState = state;
-      if (atualState is ProdutoDoCardapioCarregado) {
-        add(CarregarProdutoCardapio(atualState.produto.id));
-      } else if (state is ProdutosDoCardapioCarregados) {
-        add(CarregarProdutosCardapio());
-      }
-    });
-  }
-
-  void _pararStream(PararStreamProdutosCardapio event, Emitter<ProdutoCardapioState> emit) {
-    _listener!.cancel();
-    CardapioFirestoreServer.helper.limparDadosProdutosCardapio();
-  }
-
-  void _carregarProdutos(CarregarProdutosCardapio event, Emitter<ProdutoCardapioState> emit){
-    emit(CarregandoProdutosCardapio());
-    try {
-      emit(ProdutosDoCardapioCarregados(CardapioFirestoreServer.helper.getCardapioList()));
-    } catch (_) {
-      emit(ErroCarregarProdutosCardapio());
+  @override
+  void notificarAtualizacao(
+      NovaNotificacao event, Emitter<ProdutoCardapioState> emit) {
+    Notificacao notificacao = event.notificacao;
+    switch (notificacao.tipoNotificacao) {
+      case TipoNotificacao.conectando:
+        emit(CarregandoProdutosCardapio());
+        break;
+      case TipoNotificacao.erroConexao:
+        emit(ErroCarregarProdutosCardapio());
+        break;
+      case TipoNotificacao.novoDado:
+        ProdutoCardapioState state = this.state;
+        if (state is ProdutoDoCardapioCarregado) {
+          add(CarregarProdutoCardapio(state.produto.id));
+        } else if (state is ProdutosDoCardapioCarregados ||
+            state is CarregandoProdutosCardapio) {
+          add(CarregarProdutosCardapio());
+        }
+        break;
     }
   }
 
-  void _carregarProduto(CarregarProdutoCardapio event, Emitter<ProdutoCardapioState> emit){
+  void _carregarProdutos(
+      CarregarProdutosCardapio event, Emitter<ProdutoCardapioState> emit) {
+    //Só carrega se conexão bem sucedida.
+    //Do contrário, a stream que controla o estado da página principal
+    if (_cardapioProvider.statusConexao == StatusConexao.conectado) {
+      emit(ProdutosDoCardapioCarregados(_cardapioProvider.getCardapioList()));
+    }
+  }
+
+  void _carregarProduto(
+      CarregarProdutoCardapio event, Emitter<ProdutoCardapioState> emit) {
     emit(CarregandoProdutoCardapio());
     try {
-      ProdutoCardapio produto = CardapioFirestoreServer.helper.getProdutoCardapio(event.idProduto);
+      ProdutoCardapio produto =
+          _cardapioProvider.getProdutoCardapio(event.idProduto);
       emit(ProdutoDoCardapioCarregado(produto));
     } catch (_) {
       emit(ErroCarregarProdutoCardapio());
     }
   }
 
-  Future<void> _adicionarProduto(AdicionarAoCardapio event, Emitter<ProdutoCardapioState> emit) async {
+  Future<void> _adicionarProduto(
+      AdicionarAoCardapio event, Emitter<ProdutoCardapioState> emit) async {
     try {
-      await CardapioFirestoreServer.helper.insertProdutoCardapio(event.produto);
+      await _cardapioProvider.insertProdutoCardapio(event.produto);
       emit(SucessoAoAdicionarProdutoCardapio());
     } on FormatException catch (erro) {
       emit(ErroAoAdicionarProdutoCardapio(erro.message));
     }
   }
 
-  Future<void> _atualizarProduto(AtualizarProdutoDoCardapio event, Emitter<ProdutoCardapioState> emit) async {
+  Future<void> _atualizarProduto(AtualizarProdutoDoCardapio event,
+      Emitter<ProdutoCardapioState> emit) async {
     try {
-      await CardapioFirestoreServer.helper.updateProdutoCardapio(event.produtoAntigo, event.produtoAtual);
+      await _cardapioProvider.updateProdutoCardapio(
+          event.produtoAntigo, event.produtoAtual);
       emit(SucessoAoAtualizarProdutoCardapio());
     } on FormatException catch (erro) {
       emit(ErroAoAtualizarProdutoCardapio(erro.message));
     }
   }
 
-  Future<void> _removerProduto(RemoverDoCardapio event, Emitter<ProdutoCardapioState> emit) async {
+  Future<void> _removerProduto(
+      RemoverDoCardapio event, Emitter<ProdutoCardapioState> emit) async {
     final state = this.state;
     if (state is ProdutoDoCardapioCarregado) {
       try {
-        await CardapioFirestoreServer.helper.deleteProdutoCardapio(event.produto);
+        await _cardapioProvider.deleteProdutoCardapio(event.produto);
         emit(SucessoAoRemoverProdutoCardapio());
       } on FormatException catch (erro) {
         emit(ErroAoRemoverProdutoCardapio(erro.message));
@@ -84,7 +101,8 @@ class ProdutoCardapioBloc extends Bloc<ProdutoCardapioEvent, ProdutoCardapioStat
         emit(ProdutoDoCardapioCarregado(state.produto));
       }
     } else {
-      emit(ErroAoRemoverProdutoCardapio("Erro ao remover o produto! Caminho não estipulado"));
+      emit(ErroAoRemoverProdutoCardapio(
+          "Erro ao remover o produto! Caminho não estipulado"));
     }
   }
 }
