@@ -1,6 +1,7 @@
+import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import 'package:lanchonete/Provider/usuario_firebase_provider.dart';
+import 'package:lanchonete/provider/usuario_provider.dart';
 import 'package:lanchonete/login/bloc/autenticacao_event.dart';
 import 'package:lanchonete/login/bloc/autenticacao_state.dart';
 import 'package:lanchonete/login/models/usuario.dart';
@@ -9,9 +10,9 @@ class AutenticacaoBloc extends Bloc<AutenticacaoEvent, AutenticacaoState> {
   final UsuarioService _autenticacaoService = UsuarioService();
 
   AutenticacaoBloc() : super(Desautenticado()) {
-    _autenticacaoService.usuario.listen((event) {
-      add(EventoAutenticacaoDoServidor(event));
-    });
+    _autenticacaoService.stream.listen(
+      (event) => add(EventoAutenticacaoDoServidor(event)),
+    );
 
     on<EventoAutenticacaoDoServidor>(_emitirRespostaServidor);
     on<RegistrarUsuario>(_registrar);
@@ -22,7 +23,8 @@ class AutenticacaoBloc extends Bloc<AutenticacaoEvent, AutenticacaoState> {
     on<Deslogar>(_deslogar);
   }
 
-  _emitirRespostaServidor(EventoAutenticacaoDoServidor event, Emitter<AutenticacaoState> emit) {
+  _emitirRespostaServidor(
+      EventoAutenticacaoDoServidor event, Emitter<AutenticacaoState> emit) {
     if (event.usuario == null) {
       emit(Desautenticado());
     } else {
@@ -32,7 +34,9 @@ class AutenticacaoBloc extends Bloc<AutenticacaoEvent, AutenticacaoState> {
 
   _registrar(RegistrarUsuario event, Emitter<AutenticacaoState> emit) async {
     try {
-      await _autenticacaoService.registrar(event.email, event.senha, event.nomeLanchonete, event.endereco);
+      await _autenticacaoService.registrar(
+          event.email, event.senha, event.nomeLanchonete, event.endereco);
+      add(LogarUsuario(event.email, event.senha));
     } on FormatException catch (e) {
       emit(ErroDeAutenticacao("Erro ao Registrar!", e.message));
     }
@@ -50,17 +54,19 @@ class AutenticacaoBloc extends Bloc<AutenticacaoEvent, AutenticacaoState> {
     Usuario usuario = (state as Autenticado).usuario;
     try {
       await _autenticacaoService.deslogar();
+      emit(Desautenticado());
     } on FormatException catch (e) {
       emit(ErroDeAutenticacao("Erro ao deslogar!", e.message));
       emit(Autenticado(usuario));
     }
   }
 
-  _atualizarInformacoes(AtualizarInformacoes event, Emitter<AutenticacaoState> emit) async {
+  _atualizarInformacoes(
+      AtualizarInformacoes event, Emitter<AutenticacaoState> emit) async {
     Usuario usuario = (state as Autenticado).usuario;
     try {
-      await _autenticacaoService.atualizarInformacoes(event.nomeLanchonete, event.endereco);
-      usuario = Usuario(usuario.uid, usuario.email, event.nomeLanchonete, event.endereco);
+      await _autenticacaoService.atualizarInformacoes(
+          event.nomeLanchonete, event.endereco);
       emit(SucessoAtualizacao("Dados alterados com sucesso"));
     } on FormatException catch (e) {
       emit(ErroAtualizacao(e.message));
@@ -72,10 +78,13 @@ class AutenticacaoBloc extends Bloc<AutenticacaoEvent, AutenticacaoState> {
   _atualizarEmail(AtualizarEmail event, Emitter<AutenticacaoState> emit) async {
     Usuario usuario = (state as Autenticado).usuario;
     try {
-      await _autenticacaoService.atualizarEmailAcesso(event.senhaAtual, event.novoEmail);
+      await _autenticacaoService.atualizarEmailAcesso(
+          event.senhaAtual, event.novoEmail);
+
       emit(SucessoAtualizacao("Email alterado com sucesso"));
     } on FormatException catch (e) {
       emit(ErroAtualizacao(e.message));
+    } finally {
       emit(Autenticado(usuario));
     }
   }
@@ -83,10 +92,12 @@ class AutenticacaoBloc extends Bloc<AutenticacaoEvent, AutenticacaoState> {
   _atualizarSenha(AtualizarSenha event, Emitter<AutenticacaoState> emit) async {
     Usuario usuario = (state as Autenticado).usuario;
     try {
-      await _autenticacaoService.atualizarSenhaAcesso(event.senhaAtual, event.novaSenha);
+      await _autenticacaoService.atualizarSenhaAcesso(
+          event.senhaAtual, event.novaSenha);
       emit(SucessoAtualizacao("Senha alterada com sucesso"));
     } on FormatException catch (e) {
       emit(ErroAtualizacao(e.message));
+    } finally {
       emit(Autenticado(usuario));
     }
   }
