@@ -35,21 +35,25 @@ class TelaPedido extends StatelessWidget {
         }
 
         if (stateProdutosCardapio is ErroCarregarProdutosCardapio) {
-          return TelaErro.gerarTelaErro(_tituloDaPagina, "Erro ao carregar os produtos do cardapio para dar continuidade no pedido");
+          return TelaErro.gerarTelaErro(_tituloDaPagina,
+              "Erro ao carregar os produtos do cardapio para dar continuidade no pedido");
         }
 
         if (stateProdutosCardapio is ProdutosDoCardapioCarregados) {
           _produtosCardapio = stateProdutosCardapio.produtosCardapio;
 
           return BlocConsumer<PedidoBloc, PedidoState>(
-            buildWhen: (previous, current) => current is! EstadoDeErroPedido && current is! EstadoDeSucessoPedido,
+            buildWhen: (previous, current) =>
+                current is! EstadoDeErroPedido &&
+                current is! EstadoDeSucessoPedido,
             builder: (context, statePedido) {
               if (statePedido is CarregandoPedido) {
                 return TelaCarregamento.gerarTelaCarregamento(_tituloDaPagina);
               }
 
               if (statePedido is ErroCarregarPedido) {
-                return TelaErro.gerarTelaErro(_tituloDaPagina, "Erro ao carregar as informações do pedido");
+                return TelaErro.gerarTelaErro(_tituloDaPagina,
+                    "Erro ao carregar as informações do pedido");
               }
 
               if (statePedido is PedidoCarregado) {
@@ -59,37 +63,50 @@ class TelaPedido extends StatelessWidget {
                     actions: [
                       IconButton(
                         onPressed: () {
-                          GlobalKey<FormState> _formKey = GlobalKey<FormState>();
+                          GlobalKey<FormState> _formKey =
+                              GlobalKey<FormState>();
                           ProdutoCardapio? produtoSelecionado;
                           showDialog(
                               context: context,
                               barrierDismissible: false,
                               builder: (_) {
                                 return AlertDialog(
-                                  title: const Text('Qual o produto a ser adicionado?'),
+                                  title: const Text(
+                                      'Qual o produto a ser adicionado?'),
                                   actions: [
                                     Form(
                                       key: _formKey,
                                       child: Padding(
                                         padding: const EdgeInsets.all(8.0),
                                         child: DropdownButtonFormField<String>(
-                                          icon: const Icon(Icons.arrow_drop_down),
+                                          icon:
+                                              const Icon(Icons.arrow_drop_down),
                                           elevation: 16,
                                           isExpanded: true,
                                           onChanged: (String? un) {
-                                            produtoSelecionado = _produtosCardapio[int.parse(un!)];
+                                            produtoSelecionado =
+                                                _produtosCardapio[
+                                                    int.parse(un!)];
                                           },
-                                          items: _produtosCardapio.asMap().entries.map(
+                                          items: _produtosCardapio
+                                              .asMap()
+                                              .entries
+                                              .map(
                                             (entry) {
                                               return DropdownMenuItem<String>(
                                                 value: entry.key.toString(),
-                                                child: Text(entry.value.nomeProduto),
+                                                child: Text(
+                                                    entry.value.nomeProduto),
                                               );
                                             },
                                           ).toList(),
-                                          validator: (String? un) => un == null ? "Selecione um produto" : null,
+                                          validator: (String? un) => un == null
+                                              ? "Selecione um produto"
+                                              : null,
                                           onSaved: (String? un) {
-                                            produtoSelecionado = _produtosCardapio[int.parse(un!)];
+                                            produtoSelecionado =
+                                                _produtosCardapio[
+                                                    int.parse(un!)];
                                           },
                                         ),
                                       ),
@@ -97,20 +114,24 @@ class TelaPedido extends StatelessWidget {
                                     Padding(
                                       padding: const EdgeInsets.all(8.0),
                                       child: Row(
-                                        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                                        mainAxisAlignment:
+                                            MainAxisAlignment.spaceEvenly,
                                         children: [
                                           ElevatedButton(
-                                            onPressed: () => Navigator.pop(context, false),
+                                            onPressed: () =>
+                                                Navigator.pop(context, false),
                                             child: const Text('Cancelar'),
                                           ),
                                           ElevatedButton(
                                             onPressed: () {
-                                              if (_formKey.currentState!.validate()) {
+                                              if (_formKey.currentState!
+                                                  .validate()) {
                                                 _formKey.currentState!.save();
                                                 Navigator.pop(context, true);
                                               }
                                             },
-                                            child: const Text('Adicionar produto'),
+                                            child:
+                                                const Text('Adicionar produto'),
                                           ),
                                         ],
                                       ),
@@ -119,7 +140,10 @@ class TelaPedido extends StatelessWidget {
                                 );
                               }).then((resposta) {
                             if (resposta) {
-                              _pedidoBloc.add(AdicionarItemPedido(pedidoId, produtoSelecionado!));
+                              TelaCarregamento.gerarDialogCarregando(
+                                  context, 'Adicionando item ao pedido...');
+                              _pedidoBloc.add(AdicionarItemPedido(
+                                  pedidoId, produtoSelecionado!));
                             }
                           });
                         },
@@ -127,8 +151,12 @@ class TelaPedido extends StatelessWidget {
                       ),
                       IconButton(
                         onPressed: () {
-                          DialogConfirmacao.gerarDialogConfirmacao(context, 'Deseja realmente fechar este pedido?').then((resposta) {
+                          DialogConfirmacao.gerarDialogConfirmacao(context,
+                                  'Deseja realmente fechar este pedido?')
+                              .then((resposta) {
                             if (resposta) {
+                              TelaCarregamento.gerarDialogCarregando(
+                                  context, 'Fechando pedido...');
                               _pedidoBloc.add(FecharPedido(statePedido.pedido));
                             }
                           });
@@ -137,9 +165,14 @@ class TelaPedido extends StatelessWidget {
                       ),
                       IconButton(
                         onPressed: () {
-                          DialogConfirmacao.gerarDialogConfirmacao(context, 'Deseja realmente cancelar este pedido?').then((reposta) {
+                          DialogConfirmacao.gerarDialogConfirmacao(context,
+                                  'Deseja realmente cancelar este pedido?')
+                              .then((reposta) {
                             if (reposta) {
-                              _pedidoBloc.add(RemoverPedido(statePedido.pedido));
+                              TelaCarregamento.gerarDialogCarregando(
+                                  context, 'Cancelando pedido...');
+                              _pedidoBloc
+                                  .add(RemoverPedido(statePedido.pedido));
                             }
                           });
                         },
@@ -167,12 +200,16 @@ class TelaPedido extends StatelessWidget {
                             children: [
                               Text(
                                 'Mesa: ${statePedido.pedido.mesa}',
-                                style: const TextStyle(fontSize: 32, fontWeight: FontWeight.bold),
+                                style: const TextStyle(
+                                    fontSize: 32, fontWeight: FontWeight.bold),
                               ),
-                              const Padding(padding: EdgeInsets.symmetric(vertical: 16.0)),
+                              const Padding(
+                                  padding:
+                                      EdgeInsets.symmetric(vertical: 16.0)),
                               Text(
                                 'Valor total: R\$${statePedido.pedido.valorConta.toStringAsFixed(2)}',
-                                style: const TextStyle(fontSize: 24, fontWeight: FontWeight.w400),
+                                style: const TextStyle(
+                                    fontSize: 24, fontWeight: FontWeight.w400),
                               ),
                             ],
                           ),
@@ -184,14 +221,16 @@ class TelaPedido extends StatelessWidget {
                           decoration: BoxDecoration(color: Colors.grey[300]),
                           child: Text(
                             'Horario de abertura: ${DateFormat("HH:mm:ss dd/MM/yyyy").format(statePedido.pedido.horaAbertura)}',
-                            style: const TextStyle(fontSize: 18, fontWeight: FontWeight.w500),
+                            style: const TextStyle(
+                                fontSize: 18, fontWeight: FontWeight.w500),
                           ),
                         ),
                         const Padding(
                           padding: EdgeInsets.all(8.0),
                           child: Text(
                             'Itens pedidos:',
-                            style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
+                            style: TextStyle(
+                                fontSize: 24, fontWeight: FontWeight.bold),
                           ),
                         ),
                         Expanded(
@@ -202,25 +241,33 @@ class TelaPedido extends StatelessWidget {
                   ),
                 );
               }
-              return TelaErro.gerarTelaErro(_tituloDaPagina, "Erro ao carregar informações do pedido");
+              return TelaErro.gerarTelaErro(
+                  _tituloDaPagina, "Erro ao carregar informações do pedido");
             },
             listener: (BuildContext context, state) {
+              if (state is EstadoDeErroPedido ||
+                  state is EstadoDeSucessoPedido) {
+                Navigator.pop(context); //remove o carregamento
+              }
               if (state is EstadoDeErroPedido) {
-                NotificacaoSnackBar.gerarSnackBar(context, "ERRO: " + state.erro);
+                NotificacaoSnackBar.gerarSnackBar(
+                    context, "ERRO: " + state.erro);
               }
 
               if (state is EstadoDeSucessoPedido) {
                 NotificacaoSnackBar.gerarSnackBar(context, state.mensagem);
               }
 
-              if (state is SucessoAoFecharPedido || state is SucessoAoRemoverPedido) {
+              if (state is SucessoAoFecharPedido ||
+                  state is SucessoAoRemoverPedido) {
                 Navigator.pop(context);
               }
             },
           );
         }
 
-        return TelaErro.gerarTelaErro(_tituloDaPagina, "Erro inesperado ao carregar os produtos do cardapio.");
+        return TelaErro.gerarTelaErro(_tituloDaPagina,
+            "Erro inesperado ao carregar os produtos do cardapio.");
       },
     );
   }
@@ -248,7 +295,8 @@ class TelaPedido extends StatelessWidget {
                   Expanded(
                     child: Text(
                       item.nomeProduto,
-                      style: const TextStyle(fontWeight: FontWeight.bold, fontSize: 20.0),
+                      style: const TextStyle(
+                          fontWeight: FontWeight.bold, fontSize: 20.0),
                       maxLines: 1,
                       overflow: TextOverflow.ellipsis,
                     ),
@@ -271,12 +319,18 @@ class TelaPedido extends StatelessWidget {
                       ),
                       Container(
                         margin: const EdgeInsets.only(left: 10, top: 10),
-                        decoration: BoxDecoration(border: Border.all(width: 2), borderRadius: BorderRadius.circular(10)),
+                        decoration: BoxDecoration(
+                            border: Border.all(width: 2),
+                            borderRadius: BorderRadius.circular(10)),
                         child: Row(
                           children: [
                             IconButton(
                                 onPressed: () {
-                                  _pedidoBloc.add(AdicionarQuantidadeItemPedido(pedidoId, item, 1));
+                                  TelaCarregamento.gerarDialogCarregando(
+                                      context,
+                                      'Atualizando quantidade dos itens do pedido...');
+                                  _pedidoBloc.add(AdicionarQuantidadeItemPedido(
+                                      pedidoId, item, 1));
                                 },
                                 icon: const Icon(Icons.add)),
                             Text(
@@ -286,16 +340,26 @@ class TelaPedido extends StatelessWidget {
                             IconButton(
                               onPressed: () {
                                 if (quantidade == 1) {
-                                  DialogConfirmacao.gerarDialogConfirmacao(context, "Dejesa remover este produto dos itens pedidos?")
+                                  DialogConfirmacao.gerarDialogConfirmacao(
+                                          context,
+                                          "Dejesa remover este produto dos itens pedidos?")
                                       .then(
                                     (resposta) {
                                       if (resposta) {
-                                        _pedidoBloc.add(RemoverItemPedido(pedidoId, item));
+                                        TelaCarregamento.gerarDialogCarregando(
+                                            context,
+                                            'Removendo item do pedido...');
+                                        _pedidoBloc.add(
+                                            RemoverItemPedido(pedidoId, item));
                                       }
                                     },
                                   );
                                 } else {
-                                  _pedidoBloc.add(RemoverQuantidadeItemPedido(pedidoId, item, 1));
+                                  TelaCarregamento.gerarDialogCarregando(
+                                      context,
+                                      'Atualizando quantidade dos itens do pedido...');
+                                  _pedidoBloc.add(RemoverQuantidadeItemPedido(
+                                      pedidoId, item, 1));
                                 }
                               },
                               icon: const Icon(Icons.remove),
@@ -307,9 +371,12 @@ class TelaPedido extends StatelessWidget {
                   ),
                   TextButton.icon(
                     onPressed: () {
-                      DialogConfirmacao.gerarDialogConfirmacao(context, "Dejesa remover este produto dos itens pedidos?")
+                      DialogConfirmacao.gerarDialogConfirmacao(context,
+                              "Dejesa remover este produto dos itens pedidos?")
                           .then((resposta) {
                         if (resposta) {
+                          TelaCarregamento.gerarDialogCarregando(
+                              context, 'Removendo item do pedido...');
                           _pedidoBloc.add(RemoverItemPedido(pedidoId, item));
                         }
                       });
