@@ -12,9 +12,23 @@ class TelaErro {
     return Center(
       child: Text(
         erro,
-        style: const TextStyle(fontSize: 24,fontWeight: FontWeight.bold),
+        style: const TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
         textAlign: TextAlign.center,
       ),
+    );
+  }
+
+  static Future<dynamic> gerarDialogErro(
+      BuildContext context, String tituloErro, String mensagem) {
+    return showDialog(
+      context: context,
+      barrierDismissible: true,
+      builder: (context) {
+        return AlertDialog(
+          title: Text(tituloErro),
+          content: Text(mensagem),
+        );
+      },
     );
   }
 }
