@@ -1,16 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:lanchonete/cardapio/bloc/produto_cardapio_bloc.dart';
-import 'package:lanchonete/cardapio/bloc/produto_cardapio_event.dart';
 import 'package:lanchonete/login/bloc/autenticacao_bloc.dart';
 import 'package:lanchonete/login/bloc/autenticacao_state.dart';
 import 'package:lanchonete/login/view/tela_acesso.dart';
 import 'package:lanchonete/login/view/tela_registro.dart';
 import 'package:lanchonete/pedidos/bloc/pedido_bloc.dart';
-import 'package:lanchonete/pedidos/bloc/pedido_event.dart';
 import 'package:lanchonete/produtoEstoque/bloc/produto_estoque_bloc.dart';
-import 'package:lanchonete/produtoEstoque/bloc/produto_estoque_event.dart';
+import 'package:lanchonete/stream/bloc/stream_event.dart';
 import 'package:lanchonete/view/notificacao_snackbar.dart';
+import 'package:lanchonete/view/tela_carregamento.dart';
+import 'package:lanchonete/view/tela_erro.dart';
 import 'package:lanchonete/view/tela_principal.dart';
 
 class TelaInicial extends StatelessWidget {
@@ -21,20 +21,17 @@ class TelaInicial extends StatelessWidget {
     return BlocConsumer<AutenticacaoBloc, AutenticacaoState>(
       listener: (context, state) {
         if (state is ErroDeAutenticacao) {
-          showDialog(
-            context: context,
-            builder: (context) {
-              return AlertDialog(
-                title: Text(state.tituloErro),
-                content: Text(state.mensagem),
-              );
-            },
-          );
+          Navigator.pop(context);
+          TelaErro.gerarDialogErro(context, state.tituloErro, state.mensagem);
+          //TelaCarregamento.gerarDialogCarregando(context, 'titulo teste');
         }
         if (state is ErroAtualizacao) {
-          NotificacaoSnackBar.gerarSnackBar(context, "Erro ao atualizar: " + state.erro);
+          Navigator.pop(context);
+          NotificacaoSnackBar.gerarSnackBar(
+              context, "Erro ao atualizar: " + state.erro);
         }
         if (state is SucessoAtualizacao) {
+          Navigator.pop(context);
           NotificacaoSnackBar.gerarSnackBar(context, state.mensagem);
         }
       },
@@ -46,17 +43,21 @@ class TelaInicial extends StatelessWidget {
       builder: (context, state) {
         if (state is Autenticado) {
           if (!_estavaLogadoAnteriormente) {
-            BlocProvider.of<ProdutoEstoqueBloc>(context).add(IniciarStreamProdutosEstoque(state.usuario.uid));
-            BlocProvider.of<ProdutoCardapioBloc>(context).add(IniciarStreamProdutosCardapio(state.usuario.uid));
-            BlocProvider.of<PedidoBloc>(context).add(IniciarStreamPedidos(state.usuario.uid));
+            Navigator.pop(context);
+            BlocProvider.of<ProdutoEstoqueBloc>(context)
+                .add(IniciarStream(state.usuario.uid));
+            BlocProvider.of<ProdutoCardapioBloc>(context)
+                .add(IniciarStream(state.usuario.uid));
+            BlocProvider.of<PedidoBloc>(context)
+                .add(IniciarStream(state.usuario.uid));
             _estavaLogadoAnteriormente = true;
           }
           return TelaPrincipal();
         } else {
           if (_estavaLogadoAnteriormente) {
-            BlocProvider.of<ProdutoEstoqueBloc>(context).add(PararStreamProdutosEstoque());
-            BlocProvider.of<ProdutoCardapioBloc>(context).add(PararStreamProdutosCardapio());
-            BlocProvider.of<PedidoBloc>(context).add(PararStreamPedidos());
+            BlocProvider.of<ProdutoEstoqueBloc>(context).add(PararStream());
+            BlocProvider.of<ProdutoCardapioBloc>(context).add(PararStream());
+            BlocProvider.of<PedidoBloc>(context).add(PararStream());
             _estavaLogadoAnteriormente = false;
           }
 
